docassemble/ALWeaver/requirements.txt
docassemble.base==1.2.65
pyyaml
docx2python
